Work on #741 SP threshold control doesn't correspond to TRI2
The Grayscale representation of photon counts was 8-bit now 16-bit.  This means the probe tool will show true values.  Uses ImageProcessor.setThreshold to show a mask for the threshold.   Now you can use the same values in SP and TRI2, although the masks are a bit off (setThreshold is problematic in 16-bit mode).

Also includes some work on #744, error handling in fitted images.