WIP: Fix null displayService problem
The parameter "displayService" was already declared in
InteractiveCommand as a protected field annotated with @Parameter, which
resulted in that protected field being injected with the DisplayService,
rather than the private field in the subclass. I would consider this a
bug in ImageJ2's command framework; it should have issued you a warning
and/or error about this situation.