Update to ImgLib2
This change cleans up all the POM dependencies. It eliminates all
references to ImgLib1 classes, in favor of ImgLib2.

It also updates the versions of libraries such as ImageJ2 to the most
recent release versions, so that we can use the latest API calls.

It fixes all build errors, but there are still runtime bugs to squash.