SLIMProcessor: simplify lifetime calibration logic
The lifetime calibration can be parsed from the ImgPlus, period.
If it cannot, then it is a bug in whatever produced the ImgPlus.

And we _certainly_ do not want to parse the metadata from
format-specific hardcoded keys of the global metadata hashtable,
which is never intended to be used in such an unsafe way.