SLIMProcessor: reuse existing SciJava context
Without this change, the SLIM Curve plugin does not work: the ImgOpener
creates a new Context which attempts to create a new LegacyService (even
though it is supposed to be constrained only to SCIFIOService and
SciJavaService instances!), which throws an exception.

We should reuse the existing Context whenever possible in any case, so
this is a desirable change, although the misbehavior of ImgOpener
(probably due to a bug in SJC) will also need to be addressed.