export option with only mean values
includes changes for exporting statistics with mean values only in a CSV/TSV file. In normal execution the export option creates a very detailed file that is not very informative. This changes in this commit makes minimalistic export with only mean values of all the parameters. To disable breaif export mode, set the paramater useDetailStat in paramSetMacro.java to true. Now it is set to true