Adds the dropdown menu for seeing the mean values
This commit adds the dropdown menu in the control panel. It adds a little checkbox "Display Mean" upon selecting which user will be able to see the mean values in the dropdown menu(given the Expoprt histogram is checked). Display mean checkbox is not selected, the export will be in detail format which will inculde all possible information of the histogram. Droppdown mean values will only be visible after export and analyzing is done